requirements.txt
amqp==2.2.1
appnope==0.1.0
asn1crypto==0.22.0
attrs==17.2.0
Automat==0.6.0
Babel==2.3.4
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
backports.weakref==1.0rc1
bcrypt==3.1.3
beautifulsoup4==4.6.0
billiard==3.5.0.3
bleach==1.5.0
bson==0.4.8
celery==4.0.2
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.4
chromedriver==2.24.1
click==6.7
constantly==15.1.0
cryptography==1.9
cssselect==1.0.1
DBUtils==1.2
debtcollector==1.17.0
decorator==4.0.11
Django==1.11.2
djangorestframework==3.6.3
enum34==1.1.6
Fabric==1.13.2
fdfs-client==4.0.7
Flask==0.12.2
funcsigs==1.0.2
html5lib==0.9999999
hyperlink==17.2.1
idna==2.5
incremental==17.5.0
inflect==0.2.5
ipaddress==1.0.18
ipython==5.4.1
ipython-genutils==0.2.0
itsdangerous==0.24
Jinja2==2.9.6
kombu==4.1.0
lxml==3.8.0
Markdown==2.2.0
MarkupSafe==1.0
mock==2.0.0
mysql==0.0.1
MySQL-python==1.2.5
netaddr==0.7.19
numpy==1.13.0
olefile==0.44
oslo.config==4.11.0
oslo.i18n==3.17.0
pandas==0.20.2
paramiko==2.2.1
parsel==1.2.0
pathlib2==2.3.0
pbr==3.1.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.2.1
prompt-toolkit==1.0.14
protobuf==3.3.0
ptyprocess==0.5.1
pyasn1==0.2.3
pyasn1-modules==0.0.9
pycparser==2.18
PyDispatcher==2.0.5
Pygments==2.2.0
PyMySQL==0.7.11
PyNaCl==1.1.2
pyOpenSSL==17.1.0
PySocks==1.6.7
python-dateutil==2.6.0
pytz==2017.2
PyYAML==3.12
queuelib==1.4.2
redis==2.10.5
requesocks==0.10.8
requests==2.18.1
rfc3986==1.1.0
scandir==1.5
Scrapy==1.4.0
selenium==3.4.3
service-identity==17.0.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
sqlacodegen==1.1.6
SQLAlchemy==1.1.11
stevedore==1.25.0
tensorflow==1.2.0
threadpool==1.3.2
tornado==4.5.1
traitlets==4.3.2
tushare==0.8.2
Twisted==17.5.0
urllib3==1.21.1
vine==1.1.4
w3lib==1.17.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.12.2
wrapt==1.10.10
xlrd==1.0.0
zope.interface==4.4.2
